+++ CHANGELOG.md
@@ -11,8 +11,8 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 - Environment page.
 - "Choose your route" text.
 - Popup at first-time-visit on dashboard.
-### Changed
-- Change route link
+### Changed 
+- Footer links for display_route and comparison page
 
 ## [0.2.0] - 2020-01-13
 ### Added